% Auto-generated: do not edit by hand
\name{advancedUI}

\alias{advancedUI}

\title{AdvancedUI component}

\description{
More advanced streetscape UI.
}

\usage{
advancedUI(id=NULL, log=NULL, mapboxAccessToken=NULL, mapStyle=NULL,
xvizStyles=NULL, selectedInfo=NULL, containerStyle=NULL,
controlPanelStyle=NULL, logPanelStyle=NULL,
mapViewStyle=NULL, hudStyle=NULL, timelineStyle=NULL,
playbackControlStyle=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{log}{Named list. A string representing the logs}

\item{mapboxAccessToken}{Character. Mapbox API token}

\item{mapStyle}{Character. Map style}

\item{xvizStyles}{Named list. xviz styles}

\item{selectedInfo}{Named list. Read only prop}

\item{containerStyle}{Named list. Styling}

\item{controlPanelStyle}{Named list. Styling}

\item{logPanelStyle}{Named list. Styling}

\item{mapViewStyle}{Named list. Styling}

\item{hudStyle}{Named list. Styling}

\item{timelineStyle}{Named list. Styling}

\item{playbackControlStyle}{Named list. component style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

