% Auto-generated: do not edit by hand
\name{basicUI}

\alias{basicUI}

\title{BasicUI component}

\description{
Simple streetscape UI.
}

\usage{
basicUI(id=NULL, log=NULL, mapboxAccessToken=NULL, mapStyle=NULL,
containerStyle=NULL, logPanelStyle=NULL, mapViewStyle=NULL,
timelineStyle=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{log}{Named list. Temp}

\item{mapboxAccessToken}{Character. Mapbox API token}

\item{mapStyle}{Character. Map style}

\item{containerStyle}{Named list. Styling}

\item{logPanelStyle}{Named list. Styling}

\item{mapViewStyle}{Named list. Styling}

\item{timelineStyle}{Named list. Styling}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

